requirements.txt
google-adk>=0.3.0
google-genai>=0.4.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
httpx>=0.26.0
pydantic>=2.5.0
pypdf>=3.17.0
python-dotenv>=1.0.0
